microservices/requirements.txt
Flask==2.0.3
Flask-Cors==3.0.10
pandas==1.4.4
mysql-connector-python==8.0.33
numpy==1.23.3
prompt-toolkit==3.0.41
packaging>=21.3
Werkzeug==2.0.3
requests==2.28.1
SPARQLWrapper==2.0.0
PyYAML>=6.0
rapidfuzz>=3.5.2
